Added default values for URLs
I realized that there was a bug that if no values were supplied for the
urls, no default values were provided, and the code did not abort with
an appropriate error message.  I have added as defaults my local url for
redis development.  It works on my machine, but to do this properly the
URL's should be checked to see if they're valid, and if either of them
are not, abort and display a proper error.